Added snupkg package publish support

+++ src/Tools/CopyPackageBetweenNugetFeeds.ps1
@@ -1,15 +1,17 @@
 Param(
     [string]$version,
     [string]$server, 
-    [string]$internalServer, 
+    [string]$internalServer,
+    [string]$internalSnupkgServer,
     [string]$apiKey,
     [string]$packageId
 )
 
     $packageId = $package.Package
     $webClient = New-Object System.Net.WebClient
-    $url = "$internalServer/package/" + $packageId + "/" + $version
+    $url = "$internalServer/package/" + $packageId + "/" + $version       
     $nupkgFile = Join-Path $PSScriptRoot ($packageId + "." + $version + ".nupkg")
+
 
     Write-Host "Downloading from $url"
     $webClient.DownloadFile($url, $nupkgFile)
@@ -19,4 +21,25 @@ Param(
     & .\Tools\nuget.exe push $nupkgFile -source $server -apiKey $apiKey
     Write-Host "Package uploaded to $server."
 
-    Remove-Item $nupkgFile
+    Remove-Item $nupkgFile    
+
+    # snupkg management
+    $snupkgUrl = "file://$internalSnupkgServer/snupkg/"
+    $snupkgFile = Join-Path $PSScriptRoot ($packageId + "." + $version + ".snupkg")
+
+    try {
+      $webClient.DownloadFile($snupkgUrl, $snupkgFile)
+      $snupkgDownloaded = true;
+    }catch {
+      Write-Host "No snupkg package found!"
+      $snupkgDownloaded = false;
+   }        
+    
+    if ($snupkgDownloaded==true){
+        Write-Host "Uploading snupkg package..."        
+        & .\Tools\nuget.exe push $snupkgFile -source $server -apiKey $apiKey
+        Remove-Item $nupkgFile    
+        try {Remove-Item $snupkgFile}catch {            
+            Write-Host "Unable to cleanup snupkg..."
+        }
+    }
